[Fix/#28] - UILabel tap 이벤트 Extension 파일로 분리
